requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
elevation==1.1.3
shapely==2.0.2 